Hoist up edge display utilities to the root BaseWorkflowDisplay
