+++ ee/vellum_ee/workflows/display/types.py
@@ -7,7 +7,6 @@ from vellum.workflows.nodes import BaseNode
 from vellum.workflows.ports import Port
 from vellum.workflows.references import OutputReference, StateValueReference, WorkflowInputReference
 from vellum_ee.workflows.display.base import (
-    EdgeDisplayType,
     EntrypointDisplayType,
     StateValueDisplayType,
     WorkflowInputsDisplayType,
@@ -18,9 +17,9 @@ from vellum_ee.workflows.display.base import (
 if TYPE_CHECKING:
     from vellum_ee.workflows.display.nodes.base_node_display import BaseNodeDisplay
     from vellum_ee.workflows.display.nodes.types import NodeOutputDisplay, PortDisplay
+    from vellum_ee.workflows.display.vellum import EdgeVellumDisplay
     from vellum_ee.workflows.display.workflows import BaseWorkflowDisplay
 
-NodeDisplayType = TypeVar("NodeDisplayType", bound="BaseNodeDisplay")
 WorkflowDisplayType = TypeVar("WorkflowDisplayType", bound="BaseWorkflowDisplay")
 
 
@@ -30,9 +29,7 @@ class WorkflowDisplayContext(
         WorkflowMetaDisplayType,
         WorkflowInputsDisplayType,
         StateValueDisplayType,
-        NodeDisplayType,
         EntrypointDisplayType,
-        EdgeDisplayType,
     ]
 ):
     workflow_display_class: Type["BaseWorkflowDisplay"]
@@ -43,12 +40,12 @@ class WorkflowDisplayContext(
     )
     state_value_displays: Dict[StateValueReference, StateValueDisplayType] = field(default_factory=dict)
     global_state_value_displays: Dict[StateValueReference, StateValueDisplayType] = field(default_factory=dict)
-    node_displays: Dict[Type[BaseNode], "NodeDisplayType"] = field(default_factory=dict)
-    global_node_displays: Dict[Type[BaseNode], NodeDisplayType] = field(default_factory=dict)
+    node_displays: Dict[Type[BaseNode], "BaseNodeDisplay"] = field(default_factory=dict)
+    global_node_displays: Dict[Type[BaseNode], "BaseNodeDisplay"] = field(default_factory=dict)
     global_node_output_displays: Dict[OutputReference, Tuple[Type[BaseNode], "NodeOutputDisplay"]] = field(
         default_factory=dict
     )
     entrypoint_displays: Dict[Type[BaseNode], EntrypointDisplayType] = field(default_factory=dict)
     workflow_output_displays: Dict[BaseDescriptor, WorkflowOutputDisplay] = field(default_factory=dict)
-    edge_displays: Dict[Tuple[Port, Type[BaseNode]], EdgeDisplayType] = field(default_factory=dict)
+    edge_displays: Dict[Tuple[Port, Type[BaseNode]], "EdgeVellumDisplay"] = field(default_factory=dict)
     port_displays: Dict[Port, "PortDisplay"] = field(default_factory=dict)
